refactor(lism-ui): rename Details/Tabs internal files to Root convention (#274)
- Details.{tsx,astro} → Root.{tsx,astro}
- Tabs.{tsx,astro} → Root.{tsx,astro}
- TabList → List, TabPanel → Panel, TabItem → Item
- Update all internal import paths (external API unchanged)
