+++ packages/lism-ui/src/components/Details/astro/Root.astro
