+++ packages/lism-ui/src/components/Details/react/Root.tsx
