+++ packages/lism-ui/src/components/Details/react/index.ts
@@ -1,4 +1,4 @@
-import Root from './Details';
+import Root from './Root';
 import Summary from './Summary';
 import Title from './Title';
 import Icon from './Icon';
