+++ packages/lism-ui/src/components/Tabs/astro/List.astro
