+++ packages/lism-ui/src/components/Tabs/astro/Panel.astro
