+++ packages/lism-ui/src/components/Tabs/astro/Root.astro
@@ -4,8 +4,8 @@ import { Lism } from 'lism-css/astro';
 import atts from 'lism-css/lib/helper/atts';
 import buildModifierClass from '../../../helper/buildModifierClass';
 // import Tab from './Tab.astro';
-// import TabPanel from './TabPanel.astro';
-import TabList from './TabList.astro';
+// import TabPanel from './Panel.astro';
+import TabList from './List.astro';
 import uuid from '../../../helper/uuid.js';
 
 import transformTabitems from './transformTabitems.js';
