+++ packages/lism-ui/src/components/Tabs/astro/index.ts
@@ -1,7 +1,7 @@
-import Root from './Tabs.astro';
-import Item from './TabItem.astro';
-import List from './TabList.astro';
-import Panel from './TabPanel.astro';
+import Root from './Root.astro';
+import Item from './Item.astro';
+import List from './List.astro';
+import Panel from './Panel.astro';
 import Tab from './Tab.astro';
 
 export default { Root, List, Panel, Item, Tab };
