+++ packages/lism-ui/src/components/Tabs/react/index.ts
@@ -1,7 +1,7 @@
-import { default as Root } from './Tabs';
-import { default as Item } from './TabItem';
-import { default as List } from './TabList';
-import { default as Panel } from './TabPanel';
-import { default as Tab } from './Tab';
+import Root from './Root';
+import Item from './Item';
+import List from './List';
+import Panel from './Panel';
+import Tab from './Tab';
 
 export default { Root, List, Panel, Item, Tab };
